# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Syntellia API",
    "version": "1.0.0",
    "description": "Authenticated scan lifecycle API for creating, listing, and retrieving scan reports."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "health"
    },
    {
      "name": "scans"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["health"],
        "summary": "Service health status",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "service": { "type": "string" },
                    "ts": { "type": "string", "format": "date-time" }
                  },
                  "required": ["ok", "service", "ts"]
                }
              }
            }
          }
        }
      }
    },
    "/api/scans": {
      "get": {
        "tags": ["scans"],
        "summary": "List scans for the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["Queued", "Running", "Completed", "Failed"]
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scans fetched",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scans": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ScanRecord" }
                    },
                    "nextCursor": { "type": ["string", "null"], "format": "date-time" }
                  },
                  "required": ["scans", "nextCursor"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/AuthError" },
          "429": { "$ref": "#/components/responses/RateLimitError" },
          "500": { "$ref": "#/components/responses/ApiError" }
        }
      },
      "post": {
        "tags": ["scans"],
        "summary": "Create a new scan",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateScanInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scan created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanRecord" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApiError" },
          "401": { "$ref": "#/components/responses/AuthError" },
          "429": { "$ref": "#/components/responses/RateLimitError" },
          "500": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/scans/{scanId}": {
      "get": {
        "tags": ["scans"],
        "summary": "Fetch a single scan by id",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "scanId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanRecord" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/AuthError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/RateLimitError" },
          "500": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/scans/{scanId}/share-link": {
      "post": {
        "tags": ["scans"],
        "summary": "Create or rotate a share link for a scan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "scanId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Share link ready",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShareLinkResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/AuthError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/RateLimitError" },
          "500": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/scans/{scanId}/share-link/revoke": {
      "post": {
        "tags": ["scans"],
        "summary": "Revoke an active share link for a scan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "scanId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "Share link revoked"
          },
          "401": { "$ref": "#/components/responses/AuthError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/RateLimitError" },
          "500": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/shared/{shareToken}": {
      "get": {
        "tags": ["scans"],
        "summary": "Fetch a completed scan report by share token",
        "parameters": [
          {
            "name": "shareToken",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Shared scan found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scan": { "$ref": "#/components/schemas/SharedScanRecord" }
                  },
                  "required": ["scan"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApiError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/RateLimitError" },
          "500": { "$ref": "#/components/responses/ApiError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "AuthError": {
        "description": "Authentication failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "RateLimitError": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ApiError": {
        "description": "Request failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "CreateScanInput": {
        "type": "object",
        "properties": {
          "url": { "type": "string", "format": "uri" },
          "scanSize": {
            "type": "string",
            "enum": ["Quick check", "Standard review", "Full walkthrough"]
          },
          "loginMode": {
            "type": "string",
            "enum": ["No login needed", "This page has a login", "I'm not sure"]
          },
          "focusArea": {
            "type": "string",
            "enum": ["Overall feel", "Look and brand", "Content clarity", "Navigation and actions"]
          },
          "projectName": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": ["url"]
      },
      "ScanRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "siteName": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "scanSize": { "type": "string" },
          "loginMode": { "type": "string" },
          "focusArea": { "type": "string" },
          "projectName": { "type": "string" },
          "pageLimit": { "type": "integer", "minimum": 1 },
          "status": {
            "type": "string",
            "enum": ["Queued", "Running", "Completed", "Failed"]
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "startedAt": { "type": ["string", "null"], "format": "date-time" },
          "completedAt": { "type": ["string", "null"], "format": "date-time" },
          "error": { "type": ["string", "null"] },
          "shareToken": { "type": ["string", "null"] },
          "shareTokenExpiresAt": { "type": ["string", "null"], "format": "date-time" },
          "shareRevokedAt": { "type": ["string", "null"], "format": "date-time" },
          "report": {
            "type": ["object", "null"],
            "description": "Structured scan output with UI/styling insights (`uiStyle`) and security/technical insights (`securityTechnical`).",
            "properties": {
              "uiStyle": { "type": "object" },
              "securityTechnical": { "type": "object" }
            }
          }
        },
        "required": [
          "id",
          "userId",
          "siteName",
          "url",
          "scanSize",
          "loginMode",
          "focusArea",
          "projectName",
          "pageLimit",
          "status",
          "createdAt",
          "report"
        ]
      },
      "SharedScanRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "siteName": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "scanSize": { "type": "string" },
          "loginMode": { "type": "string" },
          "focusArea": { "type": "string" },
          "projectName": { "type": "string" },
          "pageLimit": { "type": "integer", "minimum": 1 },
          "status": {
            "type": "string",
            "enum": ["Queued", "Running", "Completed", "Failed"]
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "startedAt": { "type": ["string", "null"], "format": "date-time" },
          "completedAt": { "type": ["string", "null"], "format": "date-time" },
          "error": { "type": ["string", "null"] },
          "shareToken": { "type": ["string", "null"] },
          "shareTokenExpiresAt": { "type": ["string", "null"], "format": "date-time" },
          "report": {
            "type": ["object", "null"],
            "properties": {
              "uiStyle": { "type": "object" },
              "securityTechnical": { "type": "object" }
            }
          }
        },
        "required": [
          "id",
          "siteName",
          "url",
          "scanSize",
          "loginMode",
          "focusArea",
          "projectName",
          "pageLimit",
          "status",
          "createdAt",
          "report"
        ]
      },
      "ShareLinkResponse": {
        "type": "object",
        "properties": {
          "shareToken": { "type": "string" },
          "sharePath": { "type": "string" },
          "expiresAt": { "type": ["string", "null"], "format": "date-time" }
        },
        "required": ["shareToken", "sharePath", "expiresAt"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "requestId": { "type": "string" }
        },
        "required": ["error", "code"]
      }
    }
  }
}